DBData: move class into separate source file
